녹음 화면과 오디오 목록 화면 분리
