+++ app/src/main/AndroidManifest.xml
@@ -18,6 +18,9 @@
         android:theme="@style/Theme.SimpleRecorder"
         tools:targetApi="31">
         <activity
+            android:name=".AudioListActivity"
+            android:exported="false" />
+        <activity
             android:name=".RecordingActivity"
             android:exported="false" />
