+++ app/src/main/java/com/example/simplerecorder/RecordingState.kt
@@ -1,0 +1,7 @@
+package com.example.simplerecorder
+
+enum class RecordingState {
+    BEFORE_RECORDING,
+    ON_RECORDING,
+    PAUSE,
+}
