+++ app/src/main/res/drawable/baseline_pause_circle_filled_24.xml
@@ -1,4 +1,4 @@
-<vector android:height="40dp" android:tint="#0CB1FF"
+<vector android:height="40dp" android:tint="#F44336"
     android:viewportHeight="24" android:viewportWidth="24"
     android:width="40dp" xmlns:android="http://schemas.android.com/apk/res/android">
     <path android:fillColor="@android:color/white" android:pathData="M12,2C6.48,2 2,6.48 2,12s4.48,10 10,10 10,-4.48 10,-10S17.52,2 12,2zM11,16L9,16L9,8h2v8zM15,16h-2L13,8h2v8z"/>
