+++ app/src/main/res/drawable/baseline_stop_circle_24.xml
@@ -1,4 +1,4 @@
-<vector android:height="40dp" android:tint="#0CB1FF"
+<vector android:height="40dp" android:tint="#F44336"
     android:viewportHeight="24" android:viewportWidth="24"
     android:width="40dp" xmlns:android="http://schemas.android.com/apk/res/android">
     <path android:fillColor="@android:color/white" android:pathData="M12,2C6.48,2 2,6.48 2,12c0,5.52 4.48,10 10,10s10,-4.48 10,-10C22,6.48 17.52,2 12,2zM16,16H8V8h8V16z"/>
