+++ app/src/main/res/values/themes.xml
@@ -3,6 +3,9 @@
     <style name="Base.Theme.SimpleRecorder" parent="Theme.Material3.DayNight.NoActionBar">
         <!-- Customize your light theme here. -->
         <!-- <item name="colorPrimary">@color/my_light_primary</item> -->
+        <item name="colorPrimary">#009688</item>
+        <item name="colorPrimaryDark">#09D5C1</item>
+        <item name="android:textColorPrimary">@color/white</item>
     </style>
 
     <style name="Theme.SimpleRecorder" parent="Base.Theme.SimpleRecorder" />
